Update controller board A gerber files to v1.01
The PCB design for board A of the controller had the pins
of the Arduino socket rotated such that the USB port pointed
forward instead of towards the back of the unit.  This adds
a new revision of the board A design that corrects the
orientation of the Arduino socket.  To distinguish between
the board revisions, a v1.01 indicator has been added to the
silk screen layer.