Add reader switch control support to firmware
Most card readers have an internal switch that detects the
physical presence of a card without using the card contacts.
This switch is used by smart card reader drivers to detect
insertion and removal events.  In order to trigger these
events when we virtually switch a card, we need to control
the reader switch state from the firmware even though the
card reader adapter remains inserted.

This patch adds control of a digital output pin that is cycled
when we remove a card as well as when we switch between two card
slots.  This requires this pin to be connected to a modified
reader.  These modifications will be described in the README in
a later patch.