Update card reader modification to not include a ground wire
Connecting a ground wire between the controller and the card reader
produces a ground loop, which could potentially introduce interference.
Since the card reader and the controller are both powered by the same
computer over USB, they already share a common ground.  Thus, there is
no need for an additional ground wire.  This updates the README file
to not mention anything about adding a ground wire.