Always return an OK or ERR response for commands
Some commands don't return an OK or ERR response, as they respond
with other output (such as usage info or controller status).  This
makes it difficult for a client to know when the controller has
finished sending a multi-line response for a command.

This patch makes all commands end their response with a single
line containing either "OK" or ERR_*".  This effectively terminates
the command response.